% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badger_palette.R
\docType{data}
\name{badger_palette}
\alias{badger_palette}
\alias{badred}
\alias{badblue}
\alias{badgreen}
\alias{badgold}
\alias{badliblue}
\alias{badlightblue}
\alias{badpurple}
\title{Badger Institute color palette}
\format{
A named character vector of hexadecimal colors. The individual
  aliases are one-element character values.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
badger_palette

badred

badblue

badgreen

badgold

badliblue

badlightblue

badpurple
}
\description{
The standard six-color palette used across Badger Institute graphics. The
individual color objects are provided for compatibility with existing BTN
and chapter scripts, while `badger_palette` provides the complete named
vector.
}
\examples{
badred
badger_palette
badger_palette[c("badred", "badblue")]

}
\keyword{datasets}
