% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badger_dynamic_legend.R
\name{ggplot_add.badger_dynamic_legend}
\alias{ggplot_add.badger_dynamic_legend}
\title{Add a deferred dynamic legend to a ggplot}
\usage{
\method{ggplot_add}{badger_dynamic_legend}(object, plot, object_name)
}
\arguments{
\item{object}{A deferred dynamic-legend specification.}

\item{plot}{The ggplot assembled to the left of the `+` operator.}

\item{object_name}{Name of the object being added, supplied by ggplot2.}
}
\value{
The plot with the dynamic-legend layers added.
}
\description{
Add a deferred dynamic legend to a ggplot
}
\keyword{internal}
