% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badger_line.R
\name{ggplot_add.badger_line}
\alias{ggplot_add.badger_line}
\title{Add a deferred Badger line to a ggplot}
\usage{
\method{ggplot_add}{badger_line}(object, plot, object_name)
}
\arguments{
\item{object}{A deferred Badger-line specification.}

\item{plot}{The ggplot assembled to the left of the `+` operator.}

\item{object_name}{Name of the object being added, supplied by ggplot2.}
}
\value{
The plot with the Badger line layers added.
}
\description{
Add a deferred Badger line to a ggplot
}
\keyword{internal}
