% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Print a package greeting}
\usage{
hello()
}
\value{
The greeting string, invisibly, as returned by [print()].
}
\description{
Prints `"Hello, world!"` to the console. This function is retained for
backward compatibility with early versions of the package.
}
\examples{
hello()

}
