% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badger_finisher.R
\name{badger_finisher}
\alias{badger_finisher}
\title{Finisher}
\usage{
badger_finisher(
  plot,
  head = "Insert Headline",
  source = "Insert Source",
  logo_ref = system.file("img", "Badger-Institute-Icon.png", package = "badgerstyle"),
  filename = "plot.png",
  aspect = c("default", "1col", "2col", "web", "ppt", "custom"),
  border = T,
  height = 5,
  width = 9.55
)
}
\arguments{
\item{plot}{The ggplot2 object that the finisher will "spruce up"}

\item{head}{The headline for the plot}

\item{source}{The source attribution of the plot.}

\item{logo_ref}{The file reference for the logo.}

\item{filename}{The desired file name for the plot}

\item{aspect}{The desired aspect ratio for the plot.}
}
\description{
This function places the finishing touches on a Badger graph.
}
