% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badger_finisher.R
\name{badger_finisher}
\alias{badger_finisher}
\title{Finisher}
\usage{
badger_finisher(
  plot,
  head = "Insert Headline",
  source = "Insert Source",
  logo_ref =
    "C:/Users/Wyatt Eichholz/OneDrive - Badger Institute/R Reference/Badger-Institute-Icon.png"
)
}
\arguments{
\item{plot}{The ggplot2 object that the finisher will "spruce up"}

\item{head}{The headline for the plot}

\item{source}{The source attribution of the plot.}
}
\description{
This function places the finishing touches on a Badger graph.
}
