% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badger_publish.R
\name{badger_publish}
\alias{badger_publish}
\title{Badger Publish}
\usage{
badger_publish(filename = "plot.png", plot, aspect = c("1col", "2col", "web"))
}
\arguments{
\item{filename}{The desired filename for the image}

\item{plot}{The plot object to be saved}

\item{columns}{Whether the size parameters should be set for a one column graphic, or two}
}
\description{
This function saves the completed plot as a high quality PNG
Replaced by badger finisher.
}
