% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badger_style.R
\name{badger_style}
\alias{badger_style}
\title{Badger Style}
\usage{
badger_style(horizontal_gridlines = T, vertical_gridlines = F)
}
\description{
This creates a ggplot theme object that applies Badger Institute's style.
}
