% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_endpoints.R
\name{find_endpoints}
\alias{find_endpoints}
\title{find_endpoints}
\usage{
find_endpoints(data, x)
}
\arguments{
\item{data}{The dataframe that you will be drawing your endpoints from}

\item{x}{The variable along the x-axis.}
}
\description{
Creates an endpoint for a specified function, using ggplot commands
}
