% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_endpoint.R
\name{geom_endpoint}
\alias{geom_endpoint}
\title{geom_endpoint}
\usage{
geom_endpoint(data, x, y, color)
}
\arguments{
\item{data}{the data frame that the plot is based on}

\item{x}{the variable for the x axis}

\item{y}{the variable for the y axis}

\item{color}{the color of the endpoints}
}
\description{
This function adds endpoints a line in Badger's signature style
}
