% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_format.R
\name{num_format}
\alias{num_format}
\title{Converts}
\usage{
num_format(
  from = 1,
  to = 1,
  by = 1,
  suffix = c("k", "m", "b"),
  currency = NULL
)
}
\arguments{
\item{from}{start of range}

\item{to}{end of range}

\item{by}{interval between tick marks}

\item{suffix}{the denomination to convert, and the suffix to attack to the number}

\item{currency}{add a currency symbol to affix to the front of the value.}
}
\description{
Converts
}
