% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_format.R
\name{percent_format}
\alias{percent_format}
\title{Converts a number into a percent}
\usage{
percent_format(from = 1, to = 1, by = 1)
}
\arguments{
\item{from}{start of range}

\item{to}{end of range}

\item{by}{interval between tick marks}
}
\description{
Converts a number into a percent
}
