% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_format.R
\name{year_format}
\alias{year_format}
\title{Converts a year in the format YYYY to 'YY.}
\usage{
year_format(from = 1, to = 1, by = 1)
}
\arguments{
\item{from}{the start year}

\item{to}{the end year}

\item{by}{the interval between years}
}
\description{
Converts a year in the format YYYY to 'YY.
}
